Methods with db access and nw access made main safe
