+++ android-sdk/src/main/java/com/blueshift/core/events/BlueshiftEventRepository.kt
@@ -1,7 +1,7 @@
 package com.blueshift.core.events
 
 interface BlueshiftEventRepository {
-    fun insertEvent(event: BlueshiftEvent)
-    fun deleteEvents(events: List<BlueshiftEvent>)
-    fun readOneBatch(batchCount: Int = 100) : List<BlueshiftEvent>
+    suspend fun insertEvent(event: BlueshiftEvent)
+    suspend fun deleteEvents(events: List<BlueshiftEvent>)
+    suspend fun readOneBatch(batchCount: Int = 100) : List<BlueshiftEvent>
 }
