+++ android-sdk/src/main/java/com/blueshift/core/network/BlueshiftNetworkRepository.kt
@@ -1,5 +1,5 @@
 package com.blueshift.core.network
 
 interface BlueshiftNetworkRepository {
-    fun makeRequest(networkRequest: BlueshiftNetworkRequest) : BlueshiftNetworkResponse
+    suspend fun makeNetworkRequest(networkRequest: BlueshiftNetworkRequest) : BlueshiftNetworkResponse
 }
