+++ android-sdk/src/main/java/com/blueshift/core/network/BlueshiftNetworkRequestRepository.kt
@@ -1,8 +1,8 @@
 package com.blueshift.core.network
 
 interface BlueshiftNetworkRequestRepository {
-    fun insertRequest(networkRequest: BlueshiftNetworkRequest)
-    fun updateRequest(networkRequest: BlueshiftNetworkRequest)
-    fun deleteRequest(networkRequest: BlueshiftNetworkRequest)
-    fun readNextRequest(): BlueshiftNetworkRequest?
+    suspend fun insertRequest(networkRequest: BlueshiftNetworkRequest)
+    suspend fun updateRequest(networkRequest: BlueshiftNetworkRequest)
+    suspend fun deleteRequest(networkRequest: BlueshiftNetworkRequest)
+    suspend fun readNextRequest(): BlueshiftNetworkRequest?
 }
