+++ android-sdk/src/test/java/com/blueshift/core/events/FakeEventsRepo.kt
@@ -4,15 +4,15 @@ import kotlin.math.min
 
 class FakeEventsRepo : BlueshiftEventRepository {
      val blueshiftEvents = mutableListOf<BlueshiftEvent>()
-    override fun insertEvent(event: BlueshiftEvent) {
+    override suspend fun insertEvent(event: BlueshiftEvent) {
         blueshiftEvents.add(event)
     }
 
-    override fun deleteEvents(events: List<BlueshiftEvent>) {
+    override suspend fun deleteEvents(events: List<BlueshiftEvent>) {
         blueshiftEvents.removeAll(events)
     }
 
-    override fun readOneBatch(batchCount: Int): List<BlueshiftEvent> {
+    override suspend fun readOneBatch(batchCount: Int): List<BlueshiftEvent> {
         val result = mutableListOf<BlueshiftEvent>()
         val limit = min(batchCount, blueshiftEvents.size)
         for (i in 0 until limit) {
