+++ android-sdk/src/test/java/com/blueshift/core/network/FakeNetworkRepoWithAPIError.kt
@@ -1,7 +1,7 @@
 package com.blueshift.core.network
 
 class FakeNetworkRepoWithAPIError : BlueshiftNetworkRepository {
-    override fun makeRequest(networkRequest: BlueshiftNetworkRequest): BlueshiftNetworkResponse {
+    override suspend fun makeNetworkRequest(networkRequest: BlueshiftNetworkRequest): BlueshiftNetworkResponse {
         return BlueshiftNetworkResponse(responseCode = 500, responseBody = "{\"status\" : \"error\"}")
     }
 }
