+++ android-sdk/src/test/java/com/blueshift/core/network/FakeNetworkRepoWithAPISuccess.kt
@@ -1,7 +1,7 @@
 package com.blueshift.core.network
 
 class FakeNetworkRepoWithAPISuccess : BlueshiftNetworkRepository {
-    override fun makeRequest(networkRequest: BlueshiftNetworkRequest): BlueshiftNetworkResponse {
+    override suspend fun makeNetworkRequest(networkRequest: BlueshiftNetworkRequest): BlueshiftNetworkResponse {
         return BlueshiftNetworkResponse(responseCode = 200, responseBody = "{\"status\" : \"error\"}")
     }
 }
